# contracts/openapi.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Campaign": {
      "get": {
        "tags": [
          "Campaign"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CampaignSummaryDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CampaignSummaryDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CampaignSummaryDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Campaign"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Campaign/{id}": {
      "get": {
        "tags": [
          "Campaign"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDetailDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Campaign"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCampaignRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCampaignRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCampaignRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDetailDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Campaign"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/sharing/{contentType}/{contentId}": {
      "get": {
        "tags": [
          "Sharing"
        ],
        "parameters": [
          {
            "name": "contentType",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CustomContentType"
            }
          },
          {
            "name": "contentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContentShareDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContentShareDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContentShareDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/sharing/user": {
      "post": {
        "tags": [
          "Sharing"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareWithUserRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareWithUserRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ShareWithUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ContentShareDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentShareDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentShareDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/sharing/campaign": {
      "post": {
        "tags": [
          "Sharing"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareWithCampaignRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareWithCampaignRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ShareWithCampaignRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ContentShareDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentShareDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentShareDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/sharing/{shareId}": {
      "delete": {
        "tags": [
          "Sharing"
        ],
        "parameters": [
          {
            "name": "shareId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/sharing/{contentType}/{contentId}/visibility": {
      "put": {
        "tags": [
          "Sharing"
        ],
        "parameters": [
          {
            "name": "contentType",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CustomContentType"
            }
          },
          {
            "name": "contentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPublicRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPublicRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetPublicRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CampaignDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "dmUserId": {
            "type": "integer",
            "format": "int32"
          },
          "dmName": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "setting": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CampaignMemberDto"
            },
            "nullable": true
          },
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CampaignSessionDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CampaignMemberDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "characterId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "characterName": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "$ref": "#/components/schemas/CampaignRole"
          }
        },
        "additionalProperties": false
      },
      "CampaignRole": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
      "CampaignSessionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "sessionNumber": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "summary": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CampaignSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "dmName": {
            "type": "string",
            "nullable": true
          },
          "memberCount": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "userRole": {
            "$ref": "#/components/schemas/CampaignRole"
          }
        },
        "additionalProperties": false
      },
...
      "ContentShareDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "contentType": {
            "$ref": "#/components/schemas/CustomContentType"
          },
          "contentId": {
            "type": "integer",
            "format": "int32"
          },
          "sharedWithUserId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "sharedWithUserName": {
            "type": "string",
            "nullable": true
          },
          "sharedWithCampaignId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "sharedWithCampaignName": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "CreateCampaignRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "setting": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CustomContentType": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "SetPublicRequest": {
        "type": "object",
        "properties": {
          "isPublic": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ShareWithCampaignRequest": {
        "type": "object",
        "properties": {
          "contentType": {
            "$ref": "#/components/schemas/CustomContentType"
          },
          "contentId": {
            "type": "integer",
            "format": "int32"
          },
          "campaignId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ShareWithUserRequest": {
        "type": "object",
        "properties": {
          "contentType": {
            "$ref": "#/components/schemas/CustomContentType"
          },
          "contentId": {
            "type": "integer",
            "format": "int32"
          },
          "targetUserId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateCampaignRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "setting": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}